+++ blosc/blosclz.c
@@ -188,11 +188,6 @@ int blosclz_compress(int opt_level, const void* input, int length,
   }
   op_limit = op + maxlength;
 
-  /* output buffer cannot be less than 66 bytes or we can get into trouble */
-  if (maxlength < 66) {
-    return 0;                   /* mark this as uncompressible */
-  }
-
   htab = (uint16_t *) calloc(hash_size, sizeof(uint16_t));
 
   /* sanity check */